Remove all mentions of Nylon from README
The Nylon developer dropped the project, so Nylon probably will cease to receive updates. Now, Nylon has also caused problems for a lot of users, so its best we stop pushing for people to use it, since it causes more problems than it solves. It is also broken on Linux w/ Wine.