Hello. My name is Aspyr.
hi, i'm back again. so my last commit (that has now been sent to the void) did in fact, not fix the FC hud. turns out this whole time i had misspelled a variable, and it seems that `Change` will crash the game if it tries to modify a global that doesn't exist. so technically this isn't Aspyr's fault, but actually just the fault of Neversoft! Wow!

well, it's also my fault but you get the point.

i hope i don't have to touch this code again for a while....

Signed-off-by: kernaltrap <kernaltrap@gmail.com>